Replaces "Desc" for "Description" in types.
It is more clear and avoids issues with databases like
Postgres where using a reserved word (ie: "desc") in a table
becomes an annoying problem.

This change breaks the API but I decided to do it sooner rather than
later since the library still doesn't have a lot of adoption.